Address lints, extract list of PrefSettings to a SolverPreferencesList component
